% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_example.R
\docType{data}
\name{counts_example}
\alias{counts_example}
\title{counts_example}
\format{
\subsection{\code{counts_example}}{

A data frame with 15 rows and 4 columns:
\describe{
\item{pair}{the pair of individuals that are compared}
\item{nsnps}{the number of overlapping snps that were compared for each pair.}
\item{mismatch}{the number of sites which did not match for each pair.}
\item{pmr}{the pairwise mismatch rate (mismatch/nsnps).}
}
}
}
\usage{
counts_example
}
\description{
this is an example of the tibble made by processEigenstrat().
}
\keyword{datasets}
