% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column.R
\name{get_column_new}
\alias{get_column_new}
\title{get column}
\usage{
get_column_new(genofile, col = 1)
}
\arguments{
\item{genofile}{genofile}

\item{col}{column to return}
}
\value{
column of numbers
}
\description{
get column
}
