% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDOUGH.R
\name{plotDOUGH}
\alias{plotDOUGH}
\title{plotDOUGH}
\usage{
plotDOUGH(
  in_tibble,
  indfile,
  nsnps_cutoff = NULL,
  fntsize = 7,
  verbose = TRUE,
  removeBetween = FALSE
)
}
\arguments{
\item{in_tibble}{TBA}

\item{indfile}{TBA}

\item{nsnps_cutoff}{TBA}

\item{fntsize}{TBA}

\item{verbose}{TBA}

\item{removeBetween}{TBA}
}
\value{
plots
}
\description{
plotDOUGH
}
