% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorSensitivity.R
\name{priorSensitivity}
\alias{priorSensitivity}
\title{priorSensitivity}
\usage{
priorSensitivity(
  in_BREADR,
  row,
  degrees = NULL,
  grid_space = 0.05,
  maxPrior = 0.5
)
}
\arguments{
\item{in_BREADR}{A tibble containing the output of callRelatedness().}

\item{row}{Either the row number or pair name for which the sensitivity analysis should be run.}

\item{degrees}{A vector of integers identifying which degrees of relatedness degrees to plot. Same/Twins (1), First-Degree (2), Second-Degree (3) and Unrelated (4).}

\item{grid_space}{The space between prior probability values that are tested (smaller values mean a finer grid size).}

\item{maxPrior}{Maximum value a prior probability can take (the right-hand limit of the x-axis).}
}
\value{
plot
}
\description{
priorSensitivity
}
