% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ind.R
\name{read_ind}
\alias{read_ind}
\title{read_ind}
\usage{
read_ind(filename)
}
\arguments{
\item{filename}{a IND text file.}
}
\value{
tibble with column headings: ind (CHR), sex (CHR), pop (CHR)
}
\description{
read_ind
}
\examples{
ind_snpfile <- system.file("extdata", "example.ind.txt", package = "BREAD")
read_ind(ind_snpfile)
}
