% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_snp.R
\name{read_snp}
\alias{read_snp}
\title{read_snp}
\usage{
read_snp(filename)
}
\arguments{
\item{filename}{a SNP text file.}
}
\value{
tibble with column headings: snp (CHR), chr (DBL),
pos (DBL), site (DBL), anc (CHR), and der (CHR).
}
\description{
read_snp
}
\examples{
std_snpfile <- system.file("extdata", "example.snp.txt", package = "BREAD")
broken_snpfile <- system.file("extdata", "broken.snp.txt", package = "BREAD")
read_snp(std_snpfile)
read_snp(broken_snpfile)
}
