% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSLICES.R
\name{saveSLICES}
\alias{saveSLICES}
\title{saveSLICES}
\usage{
saveSLICES(
  in_tibble,
  outFolder = NULL,
  width = 297,
  height = 210,
  units = "mm"
)
}
\arguments{
\item{in_tibble}{a tibble that is the output of the callRelatedness() function.}

\item{outFolder}{the folder into which all diagnostic plots will be saved}

\item{width}{the width of the output PDFs.}

\item{height}{the height of the output PDFs.}

\item{units}{the units for the height and width of the output PDFs.}
}
\value{
nothing
}
\description{
Plots all pairwise diagnostic plots (in a tibble as output by callRelatedness),
as produced by plotSLICE, to a folder.
Options include the width and height of the output files, and the units in
which these dimensions are measured.
}
\examples{
\donttest{
saveSLICES(relatedness_example[1:3, ], outFolder = tempdir())
}
}
