% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_geno.R
\name{sim_geno}
\alias{sim_geno}
\title{sim_geno}
\usage{
sim_geno(n_ind, n_snp, filename)
}
\arguments{
\item{n_ind}{number of individuals}

\item{n_snp}{number of SNPs}

\item{filename}{filename of export}
}
\value{
NULL exports a file
}
\description{
Simulated geno file of eigenstrat format
}
\examples{
\dontrun{
sim_geno(10, 5, "geno.txt")
}
}
