% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_snp.R
\name{split_line}
\alias{split_line}
\title{split line}
\usage{
split_line(x)
}
\arguments{
\item{x}{line from SNP file}
}
\value{
tibble with 6 columns.
}
\description{
takes a line for a SNP file and splits into parts.
}
\examples{
split_line("1_14.570829090394763     1        0.000000              14 A X")
split_line("rs3094315	1	0.0	752566	G	A")
}
