% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDOUGH.R
\name{plotDOUGH}
\alias{plotDOUGH}
\title{plotDOUGH}
\usage{
plotDOUGH(
  in_tibble,
  indfile,
  nsnps_cutoff = NULL,
  fntsize = 7,
  verbose = TRUE,
  removeBetween = FALSE
)
}
\arguments{
\item{in_tibble}{A tibble containing the output from processEigenstrat().}

\item{indfile}{Path to an ind file with columns for the individuals IDs, genetic sex (not used) and population.}

\item{nsnps_cutoff}{Minimum number of overlapping sites for a pair to be included.}

\item{fntsize}{Font size for x axis tick labels.}

\item{verbose}{Controls the printing of progress to console.}

\item{removeBetween}{TRUE/FALSE for whether to include the “between population” PMR values as their own group (FALSE), or to exclude them (TRUE).}
}
\value{
plots
}
\description{
plotDOUGH
}
