% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorSensitivity.R
\name{priorSensitivity}
\alias{priorSensitivity}
\title{priorSensitivity}
\usage{
priorSensitivity(
  in_BREADR,
  row,
  degrees = NULL,
  grid_space = 0.05,
  maxPrior = 0.5
)
}
\arguments{
\item{in_BREADR}{TBA}

\item{row}{TBA}

\item{degrees}{TBA}

\item{grid_space}{TBA}

\item{maxPrior}{TBA}
}
\value{
plot
}
\description{
priorSensitivity
}
